requirements.txt
steam
wget
requests
Pillow
lxml
nhentai.py
pytz
hurry.filesize
Shosetsu
imgurpython
praw
cleverbot
tornado
pyyaml
discord.py
wikipedia
humanfriendly
blinker
nose
psutil
cssselect
python-twitch
wolframalpha
geopy
pymongo
google-api-python-client
git+https://github.com/abenassi/Google-Search-API
google
protobuf